rrr: migrate CompletionTracker to inline-Rust DSL (full, clean)
Replace the hand-written `class CompletionTracker` (the last non-DSL part of
completion_tracker.cpp — config/entry/status/query-result were already DSL)
with a complete inline-Rust `struct CompletionTracker` + two `#[cpp_ctor]`
factories (`new()` using CompletionTrackerConfig::defaults() for the test's
0-arg `CompletionTracker tracker_;`, and `with_config(config)`) + all ~16
methods in DSL — including the three mid-list-removal methods
(is_completed / remove / evict_expired) via index loops + `VecDeque::remove(i)`.

The LRU storage moves from `Mutex<std::list<CompletedEntry>>` to
`Mutex<VecDeque<CompletedEntry>>` (front = most recent; push_front / pop_back
preserved), so the whole class is pure-rusty with no raw iterators. Every body
is Cell/Mutex-guard/HashSet/VecDeque ops + index loops — no raw pointers,
syscalls, iterators, or operator chains.

Depends on rusty-cpp 2f1ffc8 (Mutex::new_, VecDeque::new_/remove, MutexGuard
len/contains/operator[] forwarding, cpp_ctor field-value qualification, and the
earlier inline-rust runtime-preamble suppression). Submodule bumped
c40fb2d -> 2f1ffc8. reactor.cpp's IntEvent/TimeoutEvent GEN blocks are
regenerated for the cosmetic `static_cast<>` literal-init change from the
cpp_ctor qualification fix (semantically identical).

Verified: ninja -C build_clang22 rrr links librrr.a; inline-rust --check
reports all 5 completion_tracker.cpp blocks in sync; test_completion_tracker
passes all 27 tests.

+++ reactor/reactor.cpp
@@ -215,13 +215,13 @@ struct IntEvent : public Event {
 
 IntEvent::IntEvent()
     : Event()
-    , value_(0)
-    , target_(1)
+    , value_(static_cast<int32_t>(0))
+    , target_(static_cast<int32_t>(1))
 {}
 
 IntEvent::IntEvent(int32_t tar)
     : Event()
-    , value_(0)
+    , value_(static_cast<int32_t>(0))
     , target_(tar)
 {}
